add back support for multiple documents, remove unused code
