+++ src/components/package.json
@@ -2,7 +2,6 @@
     "browser": {
         "./beginComponent.js": "./beginComponent-browser.js",
         "./endComponent.js": "./endComponent-browser.js",
-        "./event-delegation.js": "./event-delegation-browser.js",
         "./helpers.js": "./helpers-browser.js",
         "./index.js": "./index-browser.js",
         "./init-components.js": "./init-components-browser.js",
